test: clean up and consolidate test files
