+++ __tests__/tips-modal.test.tsx
@@ -344,324 +344,4 @@ describe("TipsModal", () => {
       expect(screen.getByText("Test response")).toBeTruthy();
     });
   });
-  // Tests from tips-modal-uncovered.test.tsx
-  test("prevents submission when input is empty", () => {
-    render(<TipsModal isOpen={true} onClose={mockOnClose} />);
-
-    const submitButton = screen.getByRole("button", { name: "Ask" });
-    expect(submitButton).toBeDisabled();
-
-    const form = screen
-      .getByPlaceholderText("Ask about kana learning techniques...")
-      .closest("form");
-    fireEvent.submit(form);
-
-    // fetch should not have been called
-    expect(mockFetch).not.toHaveBeenCalled();
-  });
-
-  test("prevents submission when input is only whitespace", () => {
-    render(<TipsModal isOpen={true} onClose={mockOnClose} />);
-
-    const input = screen.getByPlaceholderText(
-      "Ask about kana learning techniques...",
-    );
-    fireEvent.change(input, { target: { value: "   " } });
-
-    const submitButton = screen.getByRole("button", { name: "Ask" });
-    expect(submitButton).toBeDisabled();
-
-    const form = screen
-      .getByPlaceholderText("Ask about kana learning techniques...")
-      .closest("form");
-    fireEvent.submit(form);
-
-    // fetch should not have been called
-    expect(mockFetch).not.toHaveBeenCalled();
-  });
-
-  test("prevents submission when already loading", async () => {
-    // Mock a delayed response
-    mockFetch.mockImplementationOnce(
-      () =>
-        new Promise((resolve) => {
-          setTimeout(() => {
-            resolve(
-              mockApiResponse({
-                tip: "Test response",
-                timestamp: "2025-01-01T00:00:00Z",
-              }),
-            );
-          }, 100);
-        }),
-    );
-
-    render(<TipsModal isOpen={true} onClose={mockOnClose} />);
-
-    const input = screen.getByPlaceholderText(
-      "Ask about kana learning techniques...",
-    );
-    fireEvent.change(input, { target: { value: "First question" } });
-
-    const submitButton = screen.getByRole("button", { name: "Ask" });
-    fireEvent.click(submitButton);
-
-    // Now try to submit again while still loading
-    fireEvent.change(input, { target: { value: "Second question" } });
-    fireEvent.click(submitButton);
-
-    // fetch should only have been called once
-    expect(mockFetch).toHaveBeenCalledTimes(1);
-    expect(mockFetch).toHaveBeenCalledWith("/api/tips", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        userQuery: "First question",
-      }),
-    });
-  });
-
-  test("handles API error with custom error message", async () => {
-    const errorResponse = {
-      ok: false,
-      status: 400,
-      json: async () => ({ error: "Custom error message" }),
-    };
-
-    mockFetch.mockResolvedValue(errorResponse);
-
-    render(<TipsModal isOpen={true} onClose={mockOnClose} />);
-
-    const input = screen.getByPlaceholderText(
-      "Ask about kana learning techniques...",
-    );
-    const submitButton = screen.getByRole("button", { name: "Ask" });
-
-    fireEvent.change(input, { target: { value: "Test question" } });
-    fireEvent.click(submitButton);
-
-    await waitFor(() => {
-      expect(screen.getByText("Custom error message")).toBeInTheDocument();
-    });
-
-    // Error should be displayed in red background
-    const errorElement = screen.getByText("Custom error message");
-    expect(errorElement.closest(".bg-red-100")).toBeInTheDocument();
-  });
-
-  test("handles API error with default error message when no error is provided", async () => {
-    const errorResponse = {
-      ok: false,
-      status: 500,
-      json: async () => ({}), // No error property
-    };
-
-    mockFetch.mockResolvedValue(errorResponse);
-
-    render(<TipsModal isOpen={true} onClose={mockOnClose} />);
-
-    const input = screen.getByPlaceholderText(
-      "Ask about kana learning techniques...",
-    );
-    const submitButton = screen.getByRole("button", { name: "Ask" });
-
-    fireEvent.change(input, { target: { value: "Test question" } });
-    fireEvent.click(submitButton);
-
-    await waitFor(() => {
-      expect(
-        screen.getByText("Failed to get learning tips"),
-      ).toBeInTheDocument();
-    });
-  });
-
-  test("handles unmounted component during error processing", async () => {
-    const { unmount } = render(
-      <TipsModal isOpen={true} onClose={mockOnClose} />,
-    );
-
-    const input = screen.getByPlaceholderText(
-      "Ask about kana learning techniques...",
-    );
-    const submitButton = screen.getByRole("button", { name: "Ask" });
-
-    // Simulate component unmounting during API call
-    mockFetch.mockImplementationOnce(() => {
-      unmount(); // Unmount component while API is processing
-      return Promise.resolve({
-        ok: false,
-        status: 500,
-        json: async () => ({ error: "Test error" }),
-      });
-    });
-
-    fireEvent.change(input, { target: { value: "Test question" } });
-    fireEvent.click(submitButton);
-
-    // Wait for async operations to complete
-    await act(async () => {
-      await new Promise((resolve) => setTimeout(resolve, 100));
-    });
-
-    // Component should be unmounted without throwing errors
-    expect(screen.queryByText("Test error")).not.toBeInTheDocument();
-  });
-
-  test("handles JSON parsing error", async () => {
-    const errorResponse = {
-      ok: false,
-      status: 400,
-      json: async () => {
-        throw new Error("Invalid JSON");
-      },
-    };
-
-    mockFetch.mockResolvedValue(errorResponse);
-
-    render(<TipsModal isOpen={true} onClose={mockOnClose} />);
-
-    const input = screen.getByPlaceholderText(
-      "Ask about kana learning techniques...",
-    );
-    const submitButton = screen.getByRole("button", { name: "Ask" });
-
-    fireEvent.change(input, { target: { value: "Test question" } });
-    fireEvent.click(submitButton);
-
-    await waitFor(() => {
-      expect(screen.getByText("Invalid JSON")).toBeInTheDocument();
-    });
-  });
-
-  test("handles non-Error thrown object", async () => {
-    const errorResponse = {
-      ok: false,
-      status: 400,
-      json: async () => {
-        throw "String error";
-      },
-    };
-
-    mockFetch.mockResolvedValue(errorResponse);
-
-    render(<TipsModal isOpen={true} onClose={mockOnClose} />);
-
-    const input = screen.getByPlaceholderText(
-      "Ask about kana learning techniques...",
-    );
-    const submitButton = screen.getByRole("button", { name: "Ask" });
-
-    fireEvent.change(input, { target: { value: "Test question" } });
-    fireEvent.click(submitButton);
-
-    await waitFor(() => {
-      expect(screen.getByText("Something went wrong")).toBeInTheDocument();
-    });
-  });
-
-  test("handles null error object", async () => {
-    const errorResponse = {
-      ok: false,
-      status: 400,
-      json: async () => {
-        throw null;
-      },
-    };
-
-    mockFetch.mockResolvedValue(errorResponse);
-
-    render(<TipsModal isOpen={true} onClose={mockOnClose} />);
-
-    const input = screen.getByPlaceholderText(
-      "Ask about kana learning techniques...",
-    );
-    const submitButton = screen.getByRole("button", { name: "Ask" });
-
-    fireEvent.change(input, { target: { value: "Test question" } });
-    fireEvent.click(submitButton);
-
-    await waitFor(() => {
-      expect(screen.getByText("Something went wrong")).toBeInTheDocument();
-    });
-  });
-
-  test("handles undefined error object", async () => {
-    const errorResponse = {
-      ok: false,
-      status: 400,
-      json: async () => {
-        throw undefined;
-      },
-    };
-
-    mockFetch.mockResolvedValue(errorResponse);
-
-    render(<TipsModal isOpen={true} onClose={mockOnClose} />);
-
-    const input = screen.getByPlaceholderText(
-      "Ask about kana learning techniques...",
-    );
-    const submitButton = screen.getByRole("button", { name: "Ask" });
-
-    fireEvent.change(input, { target: { value: "Test question" } });
-    fireEvent.click(submitButton);
-
-    await waitFor(() => {
-      expect(screen.getByText("Something went wrong")).toBeInTheDocument();
-    });
-  });
-
-  test("handles number error object", async () => {
-    const errorResponse = {
-      ok: false,
-      status: 400,
-      json: async () => {
-        throw 500;
-      },
-    };
-
-    mockFetch.mockResolvedValue(errorResponse);
-
-    render(<TipsModal isOpen={true} onClose={mockOnClose} />);
-
-    const input = screen.getByPlaceholderText(
-      "Ask about kana learning techniques...",
-    );
-    const submitButton = screen.getByRole("button", { name: "Ask" });
-
-    fireEvent.change(input, { target: { value: "Test question" } });
-    fireEvent.click(submitButton);
-
-    await waitFor(() => {
-      expect(screen.getByText("Something went wrong")).toBeInTheDocument();
-    });
-  });
-
-  test("handles object error without message property", async () => {
-    const errorResponse = {
-      ok: false,
-      status: 400,
-      json: async () => {
-        throw { code: "CUSTOM_ERROR", status: 400 };
-      },
-    };
-
-    mockFetch.mockResolvedValue(errorResponse);
-
-    render(<TipsModal isOpen={true} onClose={mockOnClose} />);
-
-    const input = screen.getByPlaceholderText(
-      "Ask about kana learning techniques...",
-    );
-    const submitButton = screen.getByRole("button", { name: "Ask" });
-
-    fireEvent.change(input, { target: { value: "Test question" } });
-    fireEvent.click(submitButton);
-
-    await waitFor(() => {
-      expect(screen.getByText("Something went wrong")).toBeInTheDocument();
-    });
-  });
 });
